+++ src/FullyConnected.h
@@ -2,12 +2,12 @@
 class FullyConnected : public Layer
 {
 private:
-    uint16_t i_size, p_count;
+    uint32_t i_size, p_count;
 
 public:
-    FullyConnected(uint16_t p_c, Activation::ActivationType act, const float lr);
+    FullyConnected(uint32_t p_c, Activation::ActivationType act, const float lr);
     MNC::Matrix feed_forward(MNC::Matrix &in);
     MNC::Matrix back_propagation(const MNC::Matrix &in, const MNC::Matrix &inDer, const MNC::Matrix &err);
-    void init(uint16_t inX, uint16_t inY, uint16_t inZ);
-    void getOutDimensions(uint16_t &outX, uint16_t &outY, uint16_t &outZ);
+    void init(uint32_t inX, uint32_t inY, uint32_t inZ);
+    void getOutDimensions(uint32_t &outX, uint32_t &outY, uint32_t &outZ);
 };
