+++ src/MaxPooling.h
@@ -2,15 +2,15 @@
 class MaxPooling : public Layer
 {
 private:
-    uint16_t pooling_size;
-    uint16_t i_X, i_Y, i_Z;
-    uint16_t* inCoordX;
-    uint16_t* inCoordY;
+    uint32_t pooling_size;
+    uint32_t i_X, i_Y, i_Z;
+    uint32_t* inCoordX;
+    uint32_t* inCoordY;
 public:
-    MaxPooling(uint16_t size);
+    MaxPooling(uint32_t size);
     MNC::Matrix feed_forward(MNC::Matrix &in);
     MNC::Matrix back_propagation(const MNC::Matrix &in, const MNC::Matrix &inDer, const MNC::Matrix &err);
-    void init(uint16_t inX, uint16_t inY, uint16_t inZ);
-    void getOutDimensions(uint16_t &outX, uint16_t &outY, uint16_t &outZ);
+    void init(uint32_t inX, uint32_t inY, uint32_t inZ);
+    void getOutDimensions(uint32_t &outX, uint32_t &outY, uint32_t &outZ);
     ~MaxPooling();
 };
