+++ src/NeuralNetwork.h
@@ -6,10 +6,10 @@
 class NeuralNetwork
 {
 public:
-   uint16_t inputX, inputY, inputZ;
+   uint32_t inputX, inputY, inputZ;
    std::vector<Layer *> layers;
    void reset();
-   NeuralNetwork(uint16_t i_X, uint16_t i_Y, uint16_t i_Z);
+   NeuralNetwork(uint32_t i_X, uint32_t i_Y, uint32_t i_Z);
    ~NeuralNetwork();
    void addLayer(Layer *l);
    MNC::Matrix guess(MNC::Matrix &in);
